Initial commit: Add files, lay quite a bit of ground work
c_serv is a personal project that I want to complete within this week.
(Starting on November 9th till November 16th)... Let's go!

Essentially it should be a webserver that serves files from a particular
directory specified by the user to other computers in the network. It
should be fast (mostly) and be able to concurrently handle multiple
clients while also having a negligible learning curve to setup.

Stretch goals:
* Python flask like configuration posibilities so that with a few lines
of C code anybody can create a custom server.
* Support for post requests.