Added tests, fixed errors with general implementation
Added rudimentary tests, one to make sure the program compiles
and runs and another to make sure the thread stuff runs sanely.
Also added a small client program that is used for testing.
As of Nov 11, the server only echos whatever is sent. The next hurdle
will be to actually put in functionality to serve files. We may also
need to work on reducing CPU usage and parsing HTTP requests.

+ Also made some changes to the queue and thread architecture to get
rid of some annoying bugs that kept coming up.